requirements.txt
aiogram==3.7.0
openai==1.35.10
python-dotenv==1.0.1
httpx==0.27.0
